% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_structure_stats.R, R/performance_metrics.R
\name{compute_average_age}
\alias{compute_average_age}
\title{Compute average age of a population}
\usage{
compute_average_age(
  model_runs,
  extra_columns,
  hcr_filter,
  om_filter,
  interval_widths = c(0.5, 0.8),
  time_horizon = c(65, NA),
  extra_filter = NULL,
  relative = NULL,
  summarise_by = c("om", "hcr"),
  summary_out = TRUE
)

compute_average_age(
  model_runs,
  extra_columns,
  hcr_filter,
  om_filter,
  interval_widths = c(0.5, 0.8),
  time_horizon = c(65, NA),
  extra_filter = NULL,
  relative = NULL,
  summarise_by = c("om", "hcr"),
  summary_out = TRUE
)
}
\arguments{
\item{model_runs}{list of completed MSE simulations runs}

\item{extra_columns}{data.frame specifying names for OM and HCR to attach
to each model_run (see `bind_mse_outputs` for more details)}

\item{hcr_filter}{vector of HCR names to calculate metric over}

\item{om_filter}{vector of OM names to calculate metric over}

\item{interval_widths}{confidence intevrals to compute}

\item{extra_filter}{an additional set of filters to apply before computing 
medians and confidence intervals}

\item{relative}{a management procedure to compute metric relative to}

\item{summarise_by}{vector of columns to summarise metric by}

\item{summary_out}{whether to output data summarised by `ggdist` or full data}

\item{naa}{numbers-at-age vector}

\item{ages}{vector of ages}

\item{dem_params}{demographic parameters matrices from OM}
}
\description{
Calculate the average age of a population
given its age structure.

Compute average age (median and CIs) per year across all years and 
simulation seeds, for each combination of operating models and management 
procedures. Average age is calculated by age of individuals and is not
weighted by biomass or maturity.
}
