% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{custom_theme}
\alias{custom_theme}
\title{Custom GGPlot2 theme}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 144.
}
\usage{
custom_theme
}
\description{
Custom GGPlot2 theme
}
\keyword{datasets}
