% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_abi}
\alias{get_abi}
\title{Get Average ABI Timeseries}
\usage{
get_abi(model_runs, extra_columns, ref_naa, hcr_filter, om_filter)
}
\arguments{
\item{model_runs}{list of completed MSE simulation objects}

\item{extra_columns}{additional columns to append to output}

\item{hcr_filter}{vector of HCR names to process (must match names in `extra_columns`)}

\item{om_filter}{vector of OM names to process (must match names in `extra_columns`)}
}
\description{
Process MSE simulations for average population ABI (Griffiths et al. 2023).
}
