% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_average_age}
\alias{get_average_age}
\title{Get Average Population Age Timeseries}
\usage{
get_average_age(model_runs, extra_columns, hcr_filter, om_filter)
}
\arguments{
\item{model_runs}{list of completed MSE simulation objects}

\item{extra_columns}{additional columns to append to output}

\item{hcr_filter}{vector of HCR names to process (must match names in `extra_columns`)}

\item{om_filter}{vector of OM names to process (must match names in `extra_columns`)}
}
\description{
Process MSE simulations for average age of an individual fish
in the population. This is distinct from the performance metric
of the same name in that include both male and female fish in the
calculation of average age.
}
