% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_dynamic_economic_value}
\alias{get_dynamic_economic_value}
\title{Get Timeseries of Dynamic Economic Value}
\usage{
get_dynamic_economic_value(model_runs, extra_columns, hcr_filter, om_filter)
}
\arguments{
\item{model_runs}{list of completed MSE simulation objects}

\item{extra_columns}{additional columns to append to output}

\item{hcr_filter}{vector of HCR names to process (must match names in `extra_columns`)}

\item{om_filter}{vector of OM names to process (must match names in `extra_columns`)}
}
\description{
Process MSE simulations for dynamic economic value (Goethel et al. 2025).
Economic value calculated as a linear adjustment in price by size grade
between a max price achieved at low total landings and a min price achieved
at high total landings. Value is calculated only for the Fixed gear fleet.
}
