% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inf_min}
\alias{inf_min}
\title{Get minimum value without considering infinite values}
\usage{
inf_min(d)
}
\arguments{
\item{d}{vector of values to find maximum of}
}
\description{
Wrapper around max that ignores infinite values
}
