% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mp_25cap}
\alias{mp_25cap}
\title{An example management procedure that applies
a 25,000 mt cap on annual harvest to the 
operational F40% harvest control rule policy.}
\format{
A management procedure object.
}
\usage{
mp_25cap
}
\description{
An example management procedure that applies
a 25,000 mt cap on annual harvest to the 
operational F40% harvest control rule policy.
}
\keyword{datasets}
