% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mp_5perc}
\alias{mp_5perc}
\title{An example management procedure that applied
a symmetric 5% stability constraint to the
operational F40% harvest control rule policy.}
\format{
A management procedure object.
}
\usage{
mp_5perc
}
\description{
An example management procedure that applied
a symmetric 5% stability constraint to the
operational F40% harvest control rule policy.
}
\keyword{datasets}
