% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mp_f40}
\alias{mp_f40}
\title{An example management procedure that replicates
the NPFMC F40 harvest control rule policy that
is operationally used to manage sablefish in
Alaska.}
\format{
A management procedure object.
}
\usage{
mp_f40
}
\description{
An example management procedure that replicates
the NPFMC F40 harvest control rule policy that
is operationally used to manage sablefish in
Alaska.
}
\keyword{datasets}
