% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{om_bh_recruit}
\alias{om_bh_recruit}
\title{Beverton-Holt Recruitment OM Object}
\format{
An operating model object
}
\usage{
om_bh_recruit
}
\description{
An operating model object that will generate future
recruitment through application of a Beverton-Holt
stock recruit function.
}
\keyword{datasets}
