% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{om_bhcyclic_recruit}
\alias{om_bhcyclic_recruit}
\title{Cylic Recruitment OM Object}
\format{
An operating model object
}
\usage{
om_bhcyclic_recruit
}
\description{
An operating model object that will generate future
recruitment via a cyclic pattern of Beverton-Holt SRRs.
The first cycle is a low recruitment phase lasting 20 
years, the second cycle is a high recruitment phase
lasting 5 years.
}
\keyword{datasets}
