% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{om_crash_recruit}
\alias{om_crash_recruit}
\title{Crash Recruitment OM Object}
\format{
An operating model object
}
\usage{
om_crash_recruit
}
\description{
An operating model object that will generate future
recruitment via a recruitment crash for the first 20
years, and then resort to resampling from the historical
recruitment timeseries.
}
\keyword{datasets}
