% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{om_rand_recruit}
\alias{om_rand_recruit}
\title{Random Recruitment OM Object}
\format{
An operating model object
}
\usage{
om_rand_recruit
}
\description{
An operating model object that will generate future
recruitment by randomly resampling from teh historical
recruitment timeseries
}
\keyword{datasets}
