% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_abc_tac}
\alias{plot_abc_tac}
\title{Plot Trajectories of ABC, TAC, Expected Landings, and Attainment}
\usage{
plot_abc_tac(
  data,
  v1 = "hcr",
  v2 = NA,
  common_trajectory = 64,
  interval_widths = c(0.5, 0.8),
  base_hcr = "F40",
  highlight = NULL
)
}
\arguments{
\item{data}{tibble output from `get_management_quantities()`}

\item{v1}{variable to map to color (e.g. "hcr")}

\item{v2}{variable to facet by (e.g. "om")}

\item{common_trajectory}{year to plot vertical dashed line to indicate where MSE projection period begins}

\item{base_hcr}{HCR to plot as a thicker line for reference (must match names in `extra_columns`)}

\item{highlight}{vector of HCRs to highlight with color (must match names in `extra_columns`)}

\item{v3}{variable to facet by in addition to v2 (e.g. "om")}
}
\description{
Plot median trajectory of management quantitoes across all simulations as line plots.
}
