% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_average_age}
\alias{plot_average_age}
\title{Plot Trajectory of Average Population Age}
\usage{
plot_average_age(
  data,
  v1 = "hcr",
  v2 = NA,
  v3 = NA,
  common_trajectory = 54,
  time_horizon = NULL,
  interval_widths = c(0.5, 0.8),
  base_hcr = "F40",
  highlight = NULL,
  relative = NA
)
}
\arguments{
\item{data}{tibble output from `get_average_age()`}

\item{v1}{variable to map to color (e.g. "hcr")}

\item{v2}{variable to facet by (e.g. "om")}

\item{v3}{variable to facet by in addition to v2 (e.g. "om")}

\item{common_trajectory}{year to plot vertical dashed line to indicate where MSE projection period begins}

\item{base_hcr}{HCR to plot as a thicker line for reference (must match names in `extra_columns`)}

\item{highlight}{vector of HCRs to highlight with color (must match names in `extra_columns`)}

\item{relative}{whether to relativize average age to a specific HCR (must match names in `extra_columns`)}
}
\description{
Plot median trajectory of average population age across all simulations as a line plot.
}
