% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dynamic_economic_value}
\alias{plot_dynamic_economic_value}
\title{Plot Trajectory of Dynamic Economic Value}
\usage{
plot_dynamic_economic_value(
  data,
  v1 = "hcr",
  v2 = NA,
  v3 = NA,
  common_trajectory = 54,
  interval_widths = c(0.5, 0.8),
  base_hcr = "F40",
  highlight = NULL,
  relative = NA
)
}
\arguments{
\item{data}{tibble output from `get_dynamic_economic_value()`}

\item{v1}{variable to map to color (e.g. "hcr")}

\item{v2}{variable to facet by (e.g. "om")}

\item{v3}{variable to facet by in addition to v2 (e.g. "om")}

\item{common_trajectory}{year to plot vertical dashed line to indicate where MSE projection period begins}

\item{base_hcr}{HCR to plot as a thicker line for reference (must match names in `extra_columns`)}

\item{highlight}{vector of HCRs to highlight with color (must match names in `extra_columns`)}

\item{relative}{whether to relativize economic value to a specific HCR (must match names in `extra_columns`)}
}
\description{
Plot median trajectory of economic value for the fixed gear fleet across 
all simulations as a line plot. See `get_dynamic_economic_value()` for 
details on how economic value is calculated.
}
