% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_fishing_mortalities}
\alias{plot_fishing_mortalities}
\title{Plot Trajectory of Fishing Mortality}
\usage{
plot_fishing_mortalities(
  data,
  v1 = "hcr",
  v2 = NA,
  v3 = NA,
  show_est = FALSE,
  common_trajectory = 64,
  interval_widths = c(0.5, 0.8),
  base_hcr = "F40",
  highlight = NULL
)
}
\arguments{
\item{data}{tibble output from `get_fishing_mortalities()`}

\item{v1}{variable to map to color (e.g. "hcr")}

\item{v2}{variable to facet by (e.g. "om")}

\item{v3}{variable to facet by in addition to v2 (e.g. "om")}

\item{show_est}{whether to show estimated SSB from EM as point ranges}

\item{common_trajectory}{year to plot vertical dashed line to indicate where MSE projection period begins}

\item{base_hcr}{HCR to plot as a thicker line for reference (must match names in `extra_columns`)}

\item{highlight}{vector of HCRs to highlight with color (must match names in `extra_columns`)}
}
\description{
Plot median trajectory of total fishing mortality across all simulations as a line plot.
}
