% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_landed_catch}
\alias{plot_landed_catch}
\title{Plot Trajectory of Landed Catch}
\usage{
plot_landed_catch(
  data,
  v1 = "hcr",
  v2 = NA,
  v3 = NA,
  by_fleet = FALSE,
  common_trajectory = 64,
  base_hcr = "F40",
  highlight = NULL
)
}
\arguments{
\item{data}{tibble output from `get_landed_catch()`}

\item{v1}{variable to map to color (e.g. "hcr")}

\item{v2}{variable to facet by (e.g. "om")}

\item{v3}{variable to facet by in addition to v2 (e.g. "om")}

\item{common_trajectory}{year to plot vertical dashed line to indicate where MSE projection period begins}

\item{base_hcr}{HCR to plot as a thicker line for reference (must match names in `extra_columns`)}

\item{highlight}{vector of HCRs to highlight with color (must match names in `extra_columns`)}
}
\description{
Plot median trajectory of total landed catch across all simulations as a line plot.
}
