% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_performance_metric_summary}
\alias{plot_performance_metric_summary}
\title{Plot Summary of Performance Metrics}
\usage{
plot_performance_metric_summary(
  perf_data,
  v1 = "hcr",
  v2 = "om",
  is_relative = FALSE,
  highlight = NULL
)
}
\arguments{
\item{perf_data}{tibble output from `performance_metric_summary()`}

\item{v1}{variable to map to y-axis (e.g. "hcr")}

\item{v2}{variable to facet by (e.g. "om")}

\item{is_relative}{whether the performance metrics have been relativized to a specific HCR}

\item{highlight}{vector of HCRs to highlight with color (must match names in `extra_columns`)}
}
\description{
Plot summary of performance metrics across all simulation as point intervals.
}
