% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ssb}
\alias{plot_ssb}
\title{Plot Trajectory of Spawning Biomass}
\usage{
plot_ssb(
  data,
  v1 = "hcr",
  v2 = NA,
  v3 = NA,
  show_est = FALSE,
  common_trajectory = 64,
  base_hcr = "F40",
  relative = NA,
  highlight = NULL
)
}
\arguments{
\item{data}{tibble output from `get_ssb_biomass()`}

\item{v1}{variable to map to color (e.g. "hcr")}

\item{v2}{variable to facet by (e.g. "om")}

\item{v3}{variable to facet by in addition to v2 (e.g. "om")}

\item{show_est}{whether to show estimated SSB from EM as point ranges}

\item{common_trajectory}{year to plot vertical dashed line to indicate where MSE projection period begins}

\item{base_hcr}{HCR to plot as a thicker line for reference (must match names in `extra_columns`)}

\item{relative}{whether to relativize SSB to a specific HCR (must match names in `extra_columns`)}

\item{highlight}{vector of HCRs to highlight with color (must match names in `extra_columns`)}
}
\description{
Plot median trajectory of SSB across all simulations as a line plot.
}
