% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ssb_catch}
\alias{plot_ssb_catch}
\title{Plot Trajectories of Spawning Biomass and Landed Catch}
\usage{
plot_ssb_catch(
  ssb_data,
  catch_data,
  v1 = "hcr",
  v2 = NA,
  v3 = NA,
  common_trajectory = 64,
  base_hcr = "F40",
  flip_facet = FALSE,
  highlight = NULL
)
}
\arguments{
\item{ssb_data}{tibble output from `get_ssb_biomass()`}

\item{catch_data}{tibble output from `get_landed_catch()`}

\item{v1}{variable to map to color (e.g. "hcr")}

\item{v2}{variable to facet by (e.g. "om")}

\item{v3}{variable to facet by in addition to v2 (e.g. "om")}

\item{common_trajectory}{year to plot vertical dashed line to indicate where MSE projection period begins}

\item{base_hcr}{HCR to plot as a thicker line for reference (must match names in `extra_columns`)}

\item{highlight}{vector of HCRs to highlight with color (must match names in `extra_columns`)}
}
\description{
Plot median trajectory of total landed catch and SSB across all simulations as line plots.
}
