% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Generialize Timeseries Plotting Function}
\usage{
plot_timeseries(
  data,
  v1 = "hcr",
  v2 = NA,
  v3 = NA,
  common_trajectory = 54,
  interval_widths = c(0.5, 0.8),
  base_hcr = "F40",
  ylab = "",
  highlight = NULL
)
}
\arguments{
\item{data}{tibble summarised with `median_qi()` and reformatted with `reformat_ggdist_long()`}

\item{v1}{variable to map to color (e.g. "hcr")}

\item{v2}{variable to facet by (e.g. "om")}

\item{v3}{variable to facet by in addition to v2 (e.g. "om")}

\item{common_trajectory}{year to plot vertical dashed line to indicate where MSE projection period begins}

\item{interval_widths}{vector of credible interval widths to plot (e.g. c(0.50, 0.80))}

\item{base_hcr}{HCR to plot as a thicker line for reference (must match names in `extra_columns`)}

\item{ylab}{label for y-axis}

\item{highlight}{vector of HCRs to highlight with color (must match names in `extra_columns`)}
}
\description{
General function for plotting timeseries of management quantities (e.g. SSB, catch, ABC, TAC) 
with options for faceting and highlighting specific HCRs.
}
