% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{relativize_performance}
\alias{relativize_performance}
\title{Relativize Performance Values to a Specific Referecnce HCR}
\usage{
relativize_performance(data, rel_column, value_column, rel_value, grouping)
}
\arguments{
\item{data}{a tibble of performance metric data}

\item{rel_column}{the name of the column containing HCR names}

\item{value_column}{the name of the column containing performance metric values}

\item{rel_value}{the name of the reference HCR to relativize to}

\item{grouping}{a vector of column names to group by when relativizing (e.g. "om", "time", "sim")}
}
\description{
Convert raw performance metric values to relative performance values by dividing
by the performance metric value for a specific reference HCR. This allows for
easier comparison of performance across HCRs, as the relative performance values 
will be on a common scale.
}
