% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sablefish_assessment_data}
\alias{sablefish_assessment_data}
\title{Sablefish assessment}
\format{
A list containing assessment data
}
\usage{
sablefish_assessment_data
}
\description{
A dataset containing sablefish assessment results.
This data is used for defining the historical recruitment
timeseries.
}
\keyword{datasets}
